@@ -18,16 +18,19 @@
     "jest": "^24",
     "lerna": "^3.13",
     "typedoc": "^0.14.0",
-    "typescript": "^3.4"
+    "typescript": "^4.2"
   },
   "dependencies": {
-    "aws-sdk": "^2.7.0"
+    "@aws-sdk/client-dynamodb": "^3.10.0"
   },
   "workspaces": [
     "packages/*"
   ],
   "jest": {
     "testEnvironment": "node",
-    "testPathIgnorePatterns": ["/node_modules/", ".ts"]
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      ".ts"
+    ]
   }
 }
