@@ -18,6 +18,8 @@
   "types": "./build/index.d.ts",
   "scripts": {
     "docs": "typedoc src",
+    "build": "tsc -p tsconfig.json && tsc -p tsconfig.cjs.json",
+    "prepublish": "npm run build",
     "prepublishOnly": "tsc",
     "pretest": "tsc -p tsconfig.test.json",
     "test": "jest \"build/(.+).spec.js\""
@@ -30,13 +32,13 @@
   "devDependencies": {
     "@types/jest": "^24",
     "@types/node": "^8.0.4",
-    "aws-sdk": "^2.7.0",
+    "@aws-sdk/client-dynamodb": "^3.10.0",
     "jest": "^24",
     "typedoc": "^0.14.0",
-    "typescript": "^3.4"
+    "typescript": "^4.2"
   },
   "peerDependencies": {
-    "aws-sdk": "^2.7.0"
+    "@aws-sdk/client-dynamodb": "^3.10.0"
   },
   "dependencies": {
     "tslib": "^1.9",
