@@ -17,7 +17,8 @@
   "main": "./build/index.js",
   "types": "./build/index.d.ts",
   "scripts": {
-    "build": "tsc",
+    "build": "tsc -p tsconfig.json && tsc -p tsconfig.cjs.json",
+    "prepublish": "npm run build",
     "docs": "typedoc src",
     "integ": "npm run pretest && jest --config=jest.integration.js",
     "prepublishOnly": "npm run build",
@@ -33,10 +34,10 @@
     "@types/jest": "^24",
     "@types/node": "^8.0.4",
     "@types/uuid": "^3.0.0",
-    "aws-sdk": "^2.7.0",
+    "@aws-sdk/client-dynamodb": "^3.10.0",
     "jest": "^24",
     "typedoc": "^0.14.0",
-    "typescript": "^3.4"
+    "typescript": "^4.2"
   },
   "dependencies": {
     "@aws/dynamodb-auto-marshaller": "^0.7.1",
