@@ -2,7 +2,7 @@ import { Iterator } from './Iterator';
 import { SequentialScanOptions } from './namedParameters';
 import { ScanPaginator } from './ScanPaginator';
 import { ZeroArgumentsConstructor } from '@aws/dynamodb-data-marshaller';
-import DynamoDB = require('aws-sdk/clients/dynamodb');
+import DynamoDB = require('@aws-sdk/client-dynamodb');
 
 /**
  * Iterates over each item returned by a DynamoDB scan until no more pages are
