@@ -3,7 +3,7 @@ import { SequentialScanOptions } from './namedParameters';
 import { Paginator } from './Paginator';
 import { ScanPaginator as BasePaginator } from '@aws/dynamodb-query-iterator';
 import { ZeroArgumentsConstructor } from '@aws/dynamodb-data-marshaller';
-import DynamoDB = require('aws-sdk/clients/dynamodb');
+import DynamoDB = require('@aws-sdk/client-dynamodb');
 
 /**
  * Iterates over each page of items returned by a DynamoDB scan until no more
