@@ -4,7 +4,7 @@ import {
     ParallelScanPaginator,
     ParallelScanState,
 } from './ParallelScanPaginator';
-import DynamoDB = require('aws-sdk/clients/dynamodb');
+import {DynamoDB} from "@aws-sdk/client-dynamodb";
 
 export class ParallelScanIterator extends ItemIterator<ParallelScanPaginator> {
     constructor(
