@@ -7,7 +7,7 @@ import {
     ZeroArgumentsConstructor,
 } from '@aws/dynamodb-data-marshaller';
 import { ExpressionAttributes } from '@aws/dynamodb-expressions';
-import { ScanInput } from 'aws-sdk/clients/dynamodb';
+import { ScanInput } from '@aws-sdk/client-dynamodb';
 
 /**
  * @internal
